+++ src/hooks/index.ts
@@ -1,4 +1,19 @@
-export * from './usePromiseClient';
-export * from './useRFQ';
-export * from './useSpotPrice';
-export * from './useSoftQuote';
+export { usePromiseClient } from './usePromiseClient';
+export { type UseRFQConfig, type UseRFQReturn, useRFQ } from './useRFQ';
+export {
+  type UseSoftQuoteConfig,
+  type UseSoftQuoteReturn,
+  useSoftQuote,
+} from './useSoftQuote';
+export {
+  type UseSpotPriceConfig,
+  type UseSpotPriceReturn,
+  useSpotPrice,
+} from './useSpotPrice';
+export {
+  type MethodUnaryDescriptor,
+  createConnectQueryKey,
+  type MethodServerStreamingDescriptor,
+  type StreamResponseMessage,
+  useStream,
+} from './useStream';