+++ src/hooks/useStream.ts
@@ -1,73 +1,101 @@
-/* eslint-disable no-constant-condition -- needed for streams */
-/* eslint-disable @typescript-eslint/no-unnecessary-condition -- needed for streams */
-/* eslint-disable no-await-in-loop -- needed for streams */
-
-import type { PromiseClient } from '@connectrpc/connect';
-import { Code, ConnectError } from '@connectrpc/connect';
-import { useCallback, useEffect, useRef, useState } from 'react';
-import type { QueryClient, QueryKey } from '@tanstack/react-query';
-import { useQuery } from '@tanstack/react-query';
-import type { ServiceType } from '@bufbuild/protobuf';
+import type {
+  Message,
+  MethodInfoServerStreaming,
+  MethodInfoUnary,
+  PartialMessage,
+  ServiceType,
+} from '@bufbuild/protobuf';
+import type {
+  CallOptions,
+  ConnectError,
+  StreamResponse,
+  Transport,
+} from '@connectrpc/connect';
+import type { ConnectQueryKey } from '@connectrpc/connect-query';
+import { useTransport } from '@connectrpc/connect-query';
+import { createAsyncIterable } from '@connectrpc/connect/protocol';
+import type { UseQueryOptions } from '@tanstack/react-query';
+import { useQuery, useQueryClient } from '@tanstack/react-query';
+import { useRef, useState } from 'react';
 import { useLogger } from '../context/Logger';
 
+/** Defines a standalone method and associated service  */
+export type MethodUnaryDescriptor<
+  I extends Message<I>,
+  O extends Message<O>,
+> = MethodInfoUnary<I, O> & {
+  readonly service: Omit<ServiceType, 'methods'>;
+};
+
 /**
- * Interface for the properties accepted by the useStream hook.
+ * TanStack Query requires query keys in order to decide when the query should automatically update.
+ *
+ * In Connect-Query, much of this is handled automatically by this function.
+ *
+ * @see ConnectQueryKey for information on the components of Connect-Query's keys.
  */
-export interface UseStreamConfig<
-  TService extends ServiceType,
-  TParsedResponse,
-> {
-  /** The Query Client instance from react-query. */
-  queryClient: QueryClient;
-  /** The Query Key to use for this query in react-query. */
-  queryKey: QueryKey;
-  /** The gRPC client instance. */
-  grpcClient: PromiseClient<TService>;
-  /** The method of the gRPC client to be invoked. */
-  method: keyof PromiseClient<TService>;
-  /** The request object to send to the gRPC stream. */
-  request?: InstanceType<TService['methods'][keyof TService['methods']]['I']>;
-  /** Flag to enable or disable the stream. Defaults to true. */
-  enabled?: boolean;
-  /** The timeout in milliseconds for the gRPC request. */
-  timeoutMs?: number;
-  /** Flag to keep the stream alive even after it ends. Defaults to true. */
-  keepAlive?: boolean;
-  /** Callback function to parse responses from the gRPC stream. */
-  parseResponse?: (
-    response: InstanceType<TService['methods'][keyof TService['methods']]['O']>,
-  ) => TParsedResponse;
-  /** Callback function to handle responses from the gRPC stream. */
-  onResponse?: (
-    response: TParsedResponse extends undefined
-      ? InstanceType<TService['methods'][keyof TService['methods']]['O']>
-      : TParsedResponse,
-  ) => void;
-  /** Callback function to handle errors from the gRPC stream. */
-  onError?: (err: Error) => void;
+export function createConnectQueryKey<
+  I extends Message<I>,
+  O extends Message<O>,
+>(
+  methodDescriptor: Pick<MethodUnaryDescriptor<I, O>, 'I' | 'name' | 'service'>,
+  input: PartialMessage<I> | undefined,
+): ConnectQueryKey<I> {
+  return [
+    methodDescriptor.service.typeName,
+    methodDescriptor.name,
+    input ?? {},
+  ];
 }
 
-// Type definitions for the functions to open and abort streams.
-type OpenStreamFn = () => Promise<() => void>;
-type AbortStreamFn = () => void;
+export type MethodServerStreamingDescriptor<
+  I extends Message<I>,
+  O extends Message<O>,
+> = MethodInfoServerStreaming<I, O> & {
+  readonly service: Omit<ServiceType, 'methods'>;
+};
 
-// Defines the return type of the `useStream` hook.
-interface UseStreamReturn<TService extends ServiceType, TParsedResponse> {
-  // Array of data received from the gRPC stream.
-  data?: TParsedResponse extends undefined
-    ? InstanceType<TService['methods'][keyof TService['methods']]['O']>[]
-    : TParsedResponse[];
-  // Array of responses received from the gRPC stream.
-  responses: TParsedResponse[];
-  // Error object representing any errors that occurred during the stream.
-  error?: ConnectError | Error;
-  // Function to open a new gRPC stream.
-  openStream: OpenStreamFn;
-  // Function to abort the current gRPC stream.
-  abortStream: AbortStreamFn;
-  // Function to reset and restart the gRPC stream.
-  resetAndRestartStream: () => void;
+export interface StreamResponseMessage<T> {
+  /** List of responses in chronological order */
+  responses: T[];
+  /** Indicates if the stream is completed or not. */
+  done: boolean;
 }
+
+function handleStreamResponse<I extends Message<I>, O extends Message<O>>(
+  stream: Promise<StreamResponse<I, O>>,
+  options?: CallOptions,
+): AsyncIterable<O> {
+  // eslint-disable-next-line func-names -- generator function
+  const it = (async function* () {
+    const response = await stream;
+    options?.onHeader?.(response.header);
+    yield* response.message;
+    options?.onTrailer?.(response.trailer);
+  })()[Symbol.asyncIterator]();
+  return {
+    [Symbol.asyncIterator]: () => ({
+      next: () => it.next(),
+    }),
+  };
+}
+
+type CreateServerStreamingQueryOptions<
+  I extends Message<I>,
+  O extends Message<O>,
+  SelectOutData = StreamResponseMessage<O>,
+> = {
+  transport: Transport;
+  callOptions?: Omit<CallOptions, 'signal'> | undefined;
+} & Omit<
+  UseQueryOptions<
+    StreamResponseMessage<O>,
+    ConnectError,
+    SelectOutData,
+    ConnectQueryKey<I>
+  >,
+  'queryFn' | 'queryKey'
+>;
 
 /**
  * A React hook to manage gRPC streaming within components. It handles opening and closing streams,
@@ -76,192 +104,103 @@ interface UseStreamReturn<TService extends ServiceType, TParsedResponse> {
  * @param config - Configuration object for the gRPC streaming process.
  * @returns Object containing stream data, response handling functions, and any errors encountered.
  */
-export const useStream = <TService extends ServiceType, TParsedResponse>({
-  queryClient,
-  queryKey,
-  grpcClient,
-  method,
-  request,
-  enabled = true,
-  keepAlive = true,
-  timeoutMs,
-  parseResponse,
-  onResponse,
-  onError,
-}: UseStreamConfig<TService, TParsedResponse>): UseStreamReturn<
-  TService,
-  TParsedResponse
-> => {
+export const useStream = <
+  I extends Message<I>,
+  O extends Message<O>,
+  SelectOutData = StreamResponseMessage<O>,
+>(
+  methodSig: MethodServerStreamingDescriptor<I, O>,
+  input?: PartialMessage<I>,
+  {
+    transport,
+    callOptions,
+    onResponse,
+    timeoutMs = 15000,
+    ...queryOptions
+  }: Omit<
+    CreateServerStreamingQueryOptions<I, O, SelectOutData>,
+    'transport'
+  > & {
+    transport?: Transport;
+    onResponse?: (response: O) => void;
+    timeoutMs?: number;
+  } = {},
+) => {
+  const ctxTransport = useTransport();
+  const finalTransport = transport ?? ctxTransport;
+
+  const queryClient = useQueryClient();
+  const queryKey = createConnectQueryKey(methodSig, input);
+
   const logger = useLogger();
 
-  /**
-   * Refs are used here to keep values across renders without causing re-renders when they change.
-   */
   const streamIdRef = useRef<number>(0);
-  const abortControllerRef = useRef<AbortController | undefined>(undefined);
+  // State to track if the stream is actually open
+  const [isStreaming, setIsStreaming] = useState(false);
 
-  // State variables to hold the responses and error from the gRPC stream.
-  const [responses, setResponses] = useState<TParsedResponse[]>([]);
-  const [error, setError] = useState<ConnectError | Error>();
-
-  /**
-   * Function to abort the current stream.
-   */
-  const abortStream = useCallback(() => {
-    if (abortControllerRef.current !== undefined) {
-      logger.debug(`Aborting stream #${streamIdRef.current}`);
-      abortControllerRef.current.abort();
-      abortControllerRef.current = undefined;
-    } else if (streamIdRef.current > 0) {
-      logger.debug('Attempted to abort undefined stream.');
-    }
-  }, [logger]);
-
-  /**
-   * Function to open a new stream.
-   */
-  const openStream = useCallback(async () => {
-    // Abort any existing stream before opening a new one.
-    abortStream();
-
-    // Create a new AbortController for the upcoming stream.
-    abortControllerRef.current = new AbortController();
-    streamIdRef.current += 1;
-
-    logger.debug(`Starting stream #${streamIdRef.current}`);
-    try {
-      while (true) {
-        // @ts-expect-error never is not callable
-        for await (const res of grpcClient[method](request, {
-          signal: abortControllerRef.current.signal,
-          timeoutMs,
-        })) {
-          let response = res;
-          if (parseResponse !== undefined) {
-            try {
-              response = parseResponse(res);
-            } catch (err) {
-              onError?.(err as Error);
-            }
-          }
-          // Handle the response.
-          onResponse?.(response);
-          // Update the state with the new response.
-          setResponses((prevData) => [response, ...prevData]);
-        }
-
-        // If keepAlive is false, break out of the loop once the stream ends.
-        if (!keepAlive) break;
-      }
-    } catch (err) {
-      // Handle errors.
-      if (err instanceof Error) {
-        onError?.(err);
-        setError(err);
-      }
-
-      if (err instanceof ConnectError) {
-        const connectError = ConnectError.from(err);
-        if (connectError.code !== Code.Canceled) {
-          // Handle errors (excluding cancellations)
-          setError(connectError);
-          logger.warn({ connectError });
-        }
-      }
-    }
-
-    // Cleanup function to close the stream when the component is unmounted.
-    return () => {
-      logger.log('closing stream in cleanup');
-      abortStream();
-    };
-  }, [
-    abortStream,
-    grpcClient,
-    keepAlive,
-    logger,
-    method,
-    onError,
-    onResponse,
-    parseResponse,
-    request,
-    timeoutMs,
-  ]);
-
-  // Callback to clear the query data.
-  const clearQueryData = useCallback(
-    () => queryClient.setQueryData(queryKey, []),
-    [queryClient, queryKey],
-  );
-
-  /**
-   * Function to reset and restart the stream.
-   */
-  const resetAndRestartStream = useCallback(() => {
-    // Abort the current stream.
-    abortStream();
-    // Reset the state.
-    setResponses([]);
-    setError(undefined);
-    // Clear the query data.
-    clearQueryData();
-
-    // Restart the stream after a short delay.
-    setTimeout(() => {
-      void openStream();
-    }, 250);
-  }, [abortStream, clearQueryData, openStream]);
-
-  /**
-   * useEffect to handle the enabling and disabling of the stream based on the `enabled` prop.
-   */
-  useEffect(() => {
-    if (!enabled && abortControllerRef.current !== undefined) {
-      abortStream();
-      return;
-    }
-    if (enabled && abortControllerRef.current === undefined) {
-      void openStream();
-    } else {
-      logger.log('not opening/closing stream, but enabled just changed');
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps -- purposefully not exhaustive
-  }, [/* openStream,abortStream, */ enabled, logger]);
-
-  /**
-   * useEffect to update the query data whenever the responses state changes.
-   */
-  useEffect(() => {
-    queryClient.setQueryData(queryKey, responses);
-    // eslint-disable-next-line react-hooks/exhaustive-deps -- purposefully not exhaustive
-  }, [/* queryClient, queryKey, */ responses]);
-
-  // Use react-query to update the global query state and handle refetching.
-  const { data } = useQuery(
+  const query = useQuery({
+    ...queryOptions,
     queryKey,
-    () => {
-      if (error)
-        throw new Error(
-          error instanceof ConnectError ? error?.rawMessage : error.toString(),
-        );
-      const queryData = queryClient.getQueryData(queryKey);
-      return (queryData === undefined ? [] : queryData) as UseStreamReturn<
-        TService,
-        TParsedResponse
-      >['data'];
-    },
-    {
-      refetchInterval: 1000,
-    },
-  );
+    queryFn: async () => {
+      let responses: O[] = [];
+      const streamId = streamIdRef.current;
+      streamIdRef.current += 1;
+      logger.debug(`Starting stream #${streamId}`);
 
-  // Return values and functions for external use.
+      const abortController = new AbortController();
+      const { signal } = abortController;
+
+      // Set the timeout and save its reference
+      const timeoutId = setTimeout(() => {
+        logger.debug(`Aborting stream #${streamId}`);
+        abortController.abort();
+      }, timeoutMs);
+
+      setIsStreaming(true); // Set the stream to open when starting
+
+      try {
+        for await (const res of handleStreamResponse(
+          finalTransport.stream<I, O>(
+            {
+              typeName: methodSig.service.typeName,
+              methods: {},
+            },
+            methodSig,
+            signal,
+            callOptions?.timeoutMs,
+            callOptions?.headers,
+            createAsyncIterable([input ?? {}]),
+          ),
+        )) {
+          clearTimeout(timeoutId); // Clear the timeout upon receiving a response
+          onResponse?.(res); // Handle the response.
+          responses = [res, ...responses];
+          const newData: StreamResponseMessage<O> = {
+            done: false,
+            responses,
+          };
+          queryClient.setQueriesData({ queryKey }, newData);
+        }
+
+        return {
+          done: true,
+          responses,
+        } satisfies StreamResponseMessage<O>;
+      } catch (error) {
+        if (error instanceof Error && error.name === 'AbortError') {
+          throw new Error('Stream aborted due to timeout');
+        }
+        throw error;
+      } finally {
+        logger.log('Closing stream in cleanup...');
+        clearTimeout(timeoutId); // Ensure the timeout is cleared if the stream ends or errors
+        setIsStreaming(false); // Close the stream when it ends, errors, or aborts
+      }
+    },
+  });
+
   return {
-    data,
-    error,
-    responses,
-    openStream,
-    abortStream,
-    resetAndRestartStream,
+    ...query,
+    isStreaming,
   };
 };